Доработана функциональность по техническому заданию 3 этапа дипломного проекта

+++ core/core-common/src/main/java/ru/practicum/dto/events/EventFullDto.java
@@ -28,6 +28,6 @@ public class EventFullDto {
     private EventState state;
     private String title;
     private Long views;
-    private Long rating;
+    private Double rating;
     private ModerationCommentShortDto lastModerationCommentDto;
 }
